requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==7.0.1
exceptiongroup==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==1.1.1
nose==1.3.7
numpy==1.21.6
oauthlib==3.2.2
packaging==22.0
pandas==1.3.5
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tekore==4.5.0
tomli==2.0.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
